+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 - Added the ability to use `#[Header]`, `#[QueryString]`, and `#[RouteVariable]` to define where to get values from when generating route URIs ([#320](https://github.com/aphiria/aphiria/pull/320))
 - Added the ability to generate HTTP requests for routes with `RouteRequestFactory` ([#322](https://github.com/aphiria/aphiria/pull/322))
 - Added the ability to customize deserialization of request parameters ([#323](https://github.com/aphiria/aphiria/pull/323))
+- Added the ability to register custom route variable constraints with the `RouterComponent` ([#330](https://github.com/aphiria/aphiria/pull/330))
+- Added the ability to register custom console elements with the `CommandComponent` ([#330](https://github.com/aphiria/aphiria/pull/330))
 
 ### Fixed